refactor: update test imports to match draft command module reorganization
The draft command module has been reorganized into submodules (core, prompt_builder,
document_processor, and rag_pipeline). This commit updates all test imports and patch
paths to match the new module structure.
