--- litassist/commands/draft.py
@@ -1,524 +1,0 @@
-"""
-Citation-rich drafting via RAG & GPT-4o.
-
-This module implements the 'draft' command which uses Retrieval-Augmented Generation
-to create well-supported legal drafts. The process embeds document chunks, stores them
-in Pinecone, retrieves relevant passages using MMR re-ranking, and generates a draft
-with GPT-4o that incorporates these citations.
-"""
-
-import click
-
-from litassist.config import get_config
-from litassist.prompts import PROMPTS
-from litassist.utils.file_ops import read_document, is_text_file
-from litassist.utils.text_processing import chunk_text, create_embeddings
-from litassist.logging import save_log, save_command_output, log_task_event
-from litassist.timing import timed
-from litassist.utils.formatting import info_message
-from litassist.utils.legal_reasoning import (
-    create_reasoning_prompt,
-    detect_factual_hallucinations,
-    verify_content_if_needed,
-)
-from litassist.utils.core import show_command_completion
-from litassist.llm.factory import LLMClientFactory
-from litassist.helpers.retriever import Retriever, get_pinecone_client
-
-
-@click.command()
-@click.argument("documents", nargs=-1, required=True, type=click.Path(exists=True))
-@click.argument("query")
-@click.option(
-    "--noverify",
-    is_flag=True,
-    help="Skip standard verification",
-)
-@click.option(
-    "--diversity",
-    type=float,
-    help="Control diversity of search results (0.0-1.0)",
-    default=None,
-)
-@click.option("--output", type=str, help="Custom output filename prefix")
-@click.pass_context
-@timed
-def draft(ctx, documents, query, noverify, diversity, output):
-    """
-    Citation-rich drafting via RAG & GPT-4o.
-
-    For text files under 400KB (like case_facts.txt), passes the entire content
-    directly to the LLM for comprehensive drafting. For PDFs or larger files,
-    implements a Retrieval-Augmented Generation workflow that embeds document
-    chunks, stores them in Pinecone, and retrieves relevant passages using
-    MMR re-ranking.
-
-    Accepts multiple documents to combine knowledge from different sources
-    (e.g., case_facts.txt and strategies.txt).
-
-    Args:
-        documents: One or more paths to documents (PDF or text files) to use as knowledge base.
-                  Examples:
-                  - litassist draft case_facts.txt "query"
-                  - litassist draft case_facts.txt strategies.txt "query"
-                  - litassist draft bundle.pdf case_facts.txt "query"
-        query: The specific legal topic or argument to draft.
-        diversity: Optional float (0.0-1.0) controlling the balance between
-                  relevance and diversity in retrieved passages. Higher values
-                  prioritize diversity over relevance. (Only used for PDFs/large files)
-
-    Raises:
-        click.ClickException: If there are errors with file reading, embedding,
-                             vector storage, retrieval, or LLM API calls.
-    """
-    # Command start log
-    try:
-        log_task_event(
-            "draft",
-            "init",
-            "start",
-            "Starting draft generation",
-            {"model": LLMClientFactory.get_model_for_command("draft")},
-        )
-    except Exception:
-        pass
-
-    # Process all documents
-    try:
-        log_task_event(
-            "draft",
-            "reading",
-            "start",
-            "Reading input documents"
-        )
-    except Exception:
-        pass
-    
-    structured_content = {
-        "case_facts": "",
-        "strategies": "",
-        "other_text": [],
-        "pdf_documents": [],
-    }
-
-    for doc_path in documents:
-        text = read_document(doc_path)
-
-        # Categorize documents by type - separate file type check from size check
-        if is_text_file(doc_path):
-            # For text files, categorize by content and handle large files appropriately
-            if "case_facts" in doc_path.lower():
-                if len(text) < 400000:
-                    structured_content["case_facts"] = text
-                    click.echo(
-                        f"Using {doc_path} as CASE FACTS ({len(text)} characters)"
-                    )
-                else:
-                    # Large case facts file - use embedding/retrieval
-                    structured_content["pdf_documents"].append((doc_path, text))
-                    click.echo(f"Will use embedding/retrieval for large {doc_path}")
-            elif "strategies" in doc_path.lower() or "# Legal Strategies" in text:
-                if len(text) < 400000:
-                    structured_content["strategies"] = text
-                    click.echo(
-                        f"Using {doc_path} as LEGAL STRATEGIES ({len(text)} characters)"
-                    )
-                else:
-                    # Large strategies file - use embedding/retrieval
-                    structured_content["pdf_documents"].append((doc_path, text))
-                    click.echo(f"Will use embedding/retrieval for large {doc_path}")
-            else:
-                if len(text) < 400000:
-                    structured_content["other_text"].append((doc_path, text))
-                    click.echo(
-                        f"Using {doc_path} as supporting document ({len(text)} characters)"
-                    )
-                else:
-                    # Large text file - use embedding/retrieval
-                    structured_content["pdf_documents"].append((doc_path, text))
-                    click.echo(f"Will use embedding/retrieval for large {doc_path}")
-        else:
-            # PDF files always use embedding/retrieval
-            structured_content["pdf_documents"].append((doc_path, text))
-            click.echo(f"Will use embedding/retrieval for {doc_path}")
-
-    try:
-        log_task_event(
-            "draft",
-            "reading",
-            "end",
-            f"Read {len(documents)} document(s)"
-        )
-    except Exception:
-        pass
-    
-    # Build structured context for the LLM
-    context_parts = []
-
-    if structured_content["case_facts"]:
-        context_parts.append(
-            "=== CASE FACTS ===\n"
-            + structured_content["case_facts"]
-            + "\n=== END CASE FACTS ==="
-        )
-
-    if structured_content["strategies"]:
-        context_parts.append(
-            "=== LEGAL STRATEGIES FROM BRAINSTORMING ===\n"
-            + structured_content["strategies"]
-            + "\n=== END LEGAL STRATEGIES FROM BRAINSTORMING ==="
-        )
-
-    for doc_path, text in structured_content["other_text"]:
-        context_parts.append(
-            f"=== SUPPORTING DOCUMENT: {doc_path} ===\n{text}\n=== END SUPPORTING DOCUMENT: {doc_path} ==="
-        )
-
-    combined_text_context = "\n\n".join(context_parts) if context_parts else ""
-
-    # Process PDFs with embedding/retrieval if any
-    retrieved_context = ""
-    if structured_content["pdf_documents"]:
-        try:
-            log_task_event(
-                "draft",
-                "indexing",
-                "start",
-                "Starting document indexing for RAG pipeline"
-            )
-        except Exception:
-            pass
-        
-        # Get Pinecone client
-        pc_index = get_pinecone_client()
-
-        # Process all PDFs
-        all_chunks = []
-        doc_counter = 0
-
-        for doc_path, text in structured_content["pdf_documents"]:
-            # Chunk each document
-            chunks = chunk_text(text, max_chars=get_config().rag_max_chars)
-            for chunk in chunks:
-                doc_counter += 1
-                all_chunks.append((f"d{doc_counter}", chunk, doc_path))
-
-        # Embed all chunks
-        try:
-            log_task_event(
-                "draft",
-                "embedding",
-                "start",
-                f"Creating embeddings for {len(all_chunks)} chunks"
-            )
-        except Exception:
-            pass
-        
-        try:
-            embeddings = create_embeddings([chunk[1] for chunk in all_chunks])
-            
-            try:
-                log_task_event(
-                    "draft",
-                    "embedding",
-                    "end",
-                    f"Created {len(embeddings)} embeddings"
-                )
-            except Exception:
-                pass
-        except Exception as e:
-            raise click.ClickException(f"Embedding error: {e}")
-
-        # Create vectors with metadata
-        vectors = []
-        for i, (chunk_id, chunk_content, source_doc) in enumerate(all_chunks):
-            vectors.append(
-                (
-                    chunk_id,
-                    embeddings[i].embedding,
-                    {"text": chunk_content, "source": source_doc},
-                )
-            )
-
-        # Upsert to Pinecone
-        try:
-            log_task_event(
-                "draft",
-                "pinecone",
-                "start",
-                f"Upserting {len(vectors)} vectors to Pinecone"
-            )
-        except Exception:
-            pass
-        
-        try:
-            pc_index.upsert(vectors=vectors)
-            
-            try:
-                log_task_event(
-                    "draft",
-                    "pinecone",
-                    "end",
-                    "Vectors uploaded to Pinecone"
-                )
-            except Exception:
-                pass
-                
-        except Exception as e:
-            raise click.ClickException(f"Pinecone upsert error: {e}")
-
-        # Retrieve relevant context with MMR
-        try:
-            qemb = create_embeddings([query])[0].embedding
-        except Exception as e:
-            raise click.ClickException(f"Embedding error for query: {e}")
-
-        retriever = Retriever(pc_index, use_mmr=True)
-        
-        try:
-            log_task_event(
-                "draft",
-                "retrieval",
-                "start",
-                "Retrieving relevant context with MMR"
-            )
-        except Exception:
-            pass
-        
-        try:
-            context_list = retriever.retrieve(qemb, top_k=7, diversity_level=diversity)
-            
-            try:
-                log_task_event(
-                    "draft",
-                    "retrieval",
-                    "end",
-                    f"Retrieved {len(context_list)} relevant chunks"
-                )
-            except Exception:
-                pass
-        except Exception as e:
-            raise click.ClickException(f"Pinecone retrieval error: {e}")
-
-        retrieved_context = (
-            "\n\n=== RETRIEVED CONTEXT ===\n"
-            + "\n\n".join(context_list)
-            + "\n=== END RETRIEVED CONTEXT ==="
-        )
-        
-        try:
-            log_task_event(
-                "draft",
-                "indexing",
-                "end",
-                "RAG pipeline complete"
-            )
-        except Exception:
-            pass
-
-    # Combine all context with proper === separation
-    context = combined_text_context
-    if retrieved_context:
-        if context:
-            # Retrieved context already has its own END marker from above
-            context = context + "\n\n" + retrieved_context
-        else:
-            context = retrieved_context
-
-    # Generate draft with GPT-4o
-    client = LLMClientFactory.for_command("draft")
-    # Build system prompt based on available content
-    system_prompt = PROMPTS.get("processing.draft.system_prompt_base")
-
-    if structured_content["case_facts"] and structured_content["strategies"]:
-        system_prompt += PROMPTS.get(
-            "processing.draft.context_case_facts_and_strategies"
-        )
-    elif structured_content["case_facts"]:
-        system_prompt += PROMPTS.get("processing.draft.context_case_facts_only")
-    elif structured_content["strategies"]:
-        system_prompt += PROMPTS.get("processing.draft.context_strategies_only")
-
-    system_prompt += PROMPTS.get("processing.draft.general_instructions")
-
-    # Create user prompt using centralized template
-    user_template = PROMPTS.get("processing.draft.user_prompt_template")
-    base_user_prompt = user_template.format(
-        document_type=query, user_request=f"Context:\n{context}"
-    )
-
-    # Add reasoning trace to user prompt
-    user_prompt = create_reasoning_prompt(base_user_prompt, "draft")
-
-    messages = [
-        {"role": "system", "content": system_prompt},
-        {"role": "user", "content": user_prompt},
-    ]
-    
-    try:
-        log_task_event(
-            "draft",
-            "generation",
-            "llm_call",
-            "Sending draft generation prompt to LLM",
-            {"model": client.model}
-        )
-    except Exception:
-        pass
-    
-    try:
-        content, usage = client.complete(messages)
-        
-        try:
-            log_task_event(
-                "draft",
-                "generation",
-                "llm_response",
-                "Draft LLM response received",
-                {"model": client.model}
-            )
-        except Exception:
-            pass
-    except Exception as e:
-        raise click.ClickException(f"LLM draft error: {e}")
-
-    # Note: Citation verification now handled automatically in LLMClient.complete()
-
-    # Apply standard verification (uses verification chain like extractfacts/strategy)
-    if not noverify:
-        try:
-            log_task_event(
-                "draft",
-                "verification",
-                "start",
-                "Starting draft verification"
-            )
-        except Exception:
-            pass
-        
-        content, _ = verify_content_if_needed(
-            client, content, "draft", verify_flag=True
-        )
-        click.echo(info_message("Standard verification applied"))
-        
-        try:
-            log_task_event(
-                "draft",
-                "verification",
-                "end",
-                "Verification complete"
-            )
-        except Exception:
-            pass
-    else:
-        click.echo(info_message("Standard verification skipped by --noverify flag"))
-
-    # Track critiques for appending to output
-    critiques = []
-
-    # Check for potential hallucinations
-    try:
-        log_task_event(
-            "draft",
-            "hallucination",
-            "start",
-            "Checking for potential hallucinations"
-        )
-    except Exception:
-        pass
-    
-    hallucination_warnings = detect_factual_hallucinations(content, context)
-    
-    try:
-        log_task_event(
-            "draft",
-            "hallucination",
-            "end",
-            f"Hallucination check complete - {len(hallucination_warnings) if hallucination_warnings else 0} warnings"
-        )
-    except Exception:
-        pass
-    if hallucination_warnings:
-        # Capture hallucination warnings for critique section
-        warning_text = "The following potentially hallucinated facts were detected:\n"
-        for warning in hallucination_warnings:
-            warning_text += f"- {warning}\n"
-        warning_text += (
-            "\nPlease verify all facts against source documents before use.\n"
-        )
-        warning_text += (
-            "Replace any invented details with placeholders like [TO BE PROVIDED]."
-        )
-        critiques.append(("Factual Accuracy Warning", warning_text))
-
-        # Also add to main content for visibility
-        warning_header = "# FACTUAL ACCURACY WARNING\n\n"
-        warning_header += (
-            "The following potentially hallucinated facts were detected:\n"
-        )
-        for warning in hallucination_warnings:
-            warning_header += f"- {warning}\n"
-        warning_header += (
-            "\nPlease verify all facts against source documents before use.\n"
-        )
-        warning_header += (
-            "Replace any invented details with placeholders like [TO BE PROVIDED].\n\n"
-        )
-        warning_header += "---\n\n"
-        content = warning_header + content
-
-    # Save output using utility
-    output_file = save_command_output(
-        output if output else "draft",
-        content,
-        "" if output else query,
-        metadata={"Query": query, "Documents": ", ".join(documents)},
-        critique_sections=critiques if critiques else None,
-    )
-
-    # Reasoning trace is embedded in the main output, not saved separately
-    extra_files = None
-
-    # Save audit log (without response content)
-    save_log(
-        "draft",
-        {
-            "inputs": {
-                "documents": list(documents),
-                "query": query,
-                "context": context if context else None,
-            },
-            # Response content removed - already logged by LLMClient separately
-            "usage": usage,
-            "verification": "standard" if not noverify else "disabled",
-            "output_file": output_file,
-        },
-    )
-
-    # Show completion with preview
-    stats = {
-        "Query": query,
-        "Documents": len(documents),
-        "Verification": "Standard verification" if not noverify else "Disabled",
-    }
-
-    show_command_completion("draft", output_file, extra_files, stats)
-    
-    # Command end log
-    try:
-        log_task_event(
-            "draft",
-            "init",
-            "end",
-            "Draft generation complete"
-        )
-    except Exception:
-        pass
-
-    # Show brief preview
-    lines = content.split("\n")
-    preview_lines = [line for line in lines[:10] if line.strip()][:5]
-    if preview_lines:
-        click.echo(f"\n{info_message('Preview:')}")
-        for line in preview_lines:
-            click.echo(f"   {line[:80]}..." if len(line) > 80 else f"   {line}")
