+++ litassist/commands/draft/__init__.py
@@ -1,0 +1,4 @@
+"""Draft command package."""
+from .core import draft
+
+__all__ = ["draft"]
